# PERF: lexsort_indexer (MultiIndex / multi-column sorting)